requirements.txt
aiogram==3.13.1
pandas
openpyxl
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
